Clean up and add some database input
some clean up and addition of (# coding: utf8) i hope this will help.

I change folder name to add rpi_

i add

-air pressure to database (bmp085)
-Orientation sensor calibration status and cuz acceleration (bno055)
- more button setting like debuncing possibility

some bug correction.
