+++ src/config/main.cfg
@@ -41,35 +41,37 @@ module = plugins.test
 
 [conn_bno055]
 load = yes
-module = plugins.bno055
+module = plugins.rpi_bno055
 
 [conn_bmp085]
 load = yes
-module = plugins.bmp085
+module = plugins.rpi_bmp085
+tkey = CAT
+pkey = AIRPRESS
 
 [conn_virtualwire]
 load = yes
-module = plugins.virtualwire
+module = plugins.rpi_virtualwire
 rxpin = 23
 bps = 2000
 
 [conn_rotary_encoder]
 load = yes
-module = plugins.rotary_encoder
+module = plugins.rpi_rotary_encoder
 btn = True
-btnkey = PITCHSET
-btnstcounter = 0
-btnincr = .5
+btnkey = BARO
+btnstcounter = 29.92
+btnincr = 0.01
 btnpin = 4
 pina = 26
 pinb = 19
-stcount = 29.92
-rkey = BARO
-incr = 0.01
+stcount = 0
+rkey = PITCHSET
+incr = 1
 
 [conn_mcp3008]
 load = yes
-module = plugins.mcp3008
+module = plugins.rpi_mcp3008
 vkey1 = VOLT
 vkey2 = ANLG2
 vkey3 = ANLG3
@@ -82,6 +84,15 @@ clk = 18
 miso = 23
 mosi = 24
 cs = 25
+
+# if you want several button copie this configuration with different setting like this [conn_button2]
+[conn_button1]
+load = no
+module = plugins.rpi_button
+btnkey = BTN1
+btnpin = 4
+rdelay = 0 # 0 for debouncing or time in seconde to determine the repeat delay
+
 
 # Logging configuration - See Python logging.config module documenation
 [loggers]
