+++ src/plugins/rpi_bmp085/Adafruit_BMP/BMP085.py
@@ -1,3 +1,4 @@
+# coding: utf8
 # Copyright (c) 2014 Adafruit Industries
 # Author: Tony DiCola
 #
