+++ src/plugins/rpi_bno055/Adafruit_BNO055/BNO055.py
@@ -1,3 +1,4 @@
+# coding: utf8
 # Adafruit BNO055 Absolute Orientation Sensor Library
 # Copyright (c) 2015 Adafruit Industries
 # Author: Tony DiCola
