+++ src/plugins/rpi_mcp3008/Adafruit_MCP3008/MCP3008.py
@@ -1,3 +1,4 @@
+# coding: utf8
 # Copyright (c) 2016 Adafruit Industries
 # Author: Tony DiCola
 #
