+++ src/plugins/rpi_mcp3008/Adafruit_MCP3008/__init__.py
