+++ src/plugins/rpi_rotary_encoder/__init__.py
@@ -1,6 +1,7 @@
+# coding: utf8
 #!/usr/bin/env python
 
-#  Copyright (c) 2014 Phil Birkelbach
+#  Copyright (c) 2017 Jean-Manuel Gagnon
 #
 #  This program is free software; you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
@@ -24,10 +25,9 @@
 import plugin
 import threading
 import time
-import Adafruit_GPIO.SPI as SPI
-import Adafruit_MCP3008
+from RPi import GPIO
 from collections import OrderedDict
-
+# TODO create a menu driver between setting, like Garmin gns or gnc serie
 class MainThread(threading.Thread):
     def __init__(self, parent):
         """The calling object should pass itself as the parent.
@@ -37,47 +37,47 @@ class MainThread(threading.Thread):
         self.getout = False   # indicator for when to stop
         self.parent = parent  # parent plugin object
         self.log = parent.log  # simplifies logging
+        self.key = parent.config['rkey'] if ('rkey' in parent.config) and parent.config['rkey'] else "ENC1"
+        self.btn = parent.config['btn'] if ('btn' in parent.config) and parent.config['btn'] else "False"
+        self.btnpin = int(parent.config['btnpin']) if ('btnpin' in parent.config) and parent.config['btnpin'] else 4
+        self.btnkey = parent.config['btnkey'] if ('btnkey' in parent.config) and parent.config['btnkey'] else "BTN1"
+        self.btnstcounter = float(parent.config['btnstcounter']) if ('btnstcounter' in parent.config) and parent.config['btnstcounter'] else 0
+        self.btnincr = float(parent.config['btnincr']) if ('btnincr' in parent.config) and parent.config['btnincr'] else 1
+        self.counter = float(parent.config['stcount']) if ('stcount' in parent.config) and parent.config['stcount'] else 0
+        self.incr = float(parent.config['incr']) if ('incr' in parent.config) and parent.config['incr'] else 1
+        self.clk = int(parent.config['pina']) if ('pina' in parent.config) and parent.config['pina'] else 19
+        self.dt = int(parent.config['pinb']) if ('pinb' in parent.config) and parent.config['pinb'] else 26
+        GPIO.setmode(GPIO.BCM)
+        GPIO.setup(self.btnpin,GPIO.IN, pull_up_down=GPIO.PUD_UP)
+        GPIO.setup(self.clk, GPIO.IN, pull_up_down=GPIO.PUD_DOWN)
+        GPIO.setup(self.dt, GPIO.IN, pull_up_down=GPIO.PUD_DOWN)
         self.count = 0
-        SPI_PORT   = 0
-        SPI_DEVICE = 0
-        self.CLK = int(parent.config['clk']) if ('clk' in parent.config) and parent.config['clk'] else 18
-        self.MISO = int(parent.config['miso']) if ('miso' in parent.config) and parent.config['miso'] else 23
-        self.MOSI = int(parent.config['mosi']) if ('mosi' in parent.config) and parent.config['mosi'] else 24
-        self.CS = int(parent.config['cs']) if ('cs' in parent.config) and parent.config['cs'] else 25
-	self.VKEY1 = parent.config['vkey1'] if ('vkey1' in parent.config) and parent.config['vkey1'] else "ANLG1"
-	self.VKEY2 = parent.config['vkey2'] if ('vkey2' in parent.config) and parent.config['vkey2'] else "ANLG2"
-	self.VKEY3 = parent.config['vkey3'] if ('vkey3' in parent.config) and parent.config['vkey3'] else "ANLG3"
-	self.VKEY4 = parent.config['vkey4'] if ('vkey4' in parent.config) and parent.config['vkey4'] else "ANLG4"
-	self.VKEY5 = parent.config['vkey5'] if ('vkey5' in parent.config) and parent.config['vkey5'] else "ANLG5"
-	self.VKEY6 = parent.config['vkey6'] if ('vkey6' in parent.config) and parent.config['vkey6'] else "ANLG6"
-	self.VKEY7 = parent.config['vkey7'] if ('vkey7' in parent.config) and parent.config['vkey7'] else "ANLG7"
-	self.VKEY8 = parent.config['vkey8'] if ('vkey8' in parent.config) and parent.config['vkey8'] else "ANLG8"
-
-	#self.mcp = Adafruit_MCP3008.MCP3008(clk=self.CLK, cs=self.CS, miso=self.MISO, mosi=self.MOSI)
-	self.mcp = Adafruit_MCP3008.MCP3008(spi=SPI.SpiDev(SPI_PORT, SPI_DEVICE))
+        self.clkLastState = GPIO.input(self.clk)
 
     def run(self):
         while True:
             if self.getout:
                 break
-            time.sleep(1)
+            time.sleep(0.001)
             self.count += 1
-            value1 = self.mcp.read_adc(0)
-            value2 = self.mcp.read_adc(1)
-            value3 = self.mcp.read_adc(2)
-            value4 = self.mcp.read_adc(3)
-            value5 = self.mcp.read_adc(4)
-            value6 = self.mcp.read_adc(5)
-            value7 = self.mcp.read_adc(6)
-            value8 = self.mcp.read_adc(7)
-            self.parent.db_write(self.VKEY1, value1)
-            self.parent.db_write(self.VKEY2, value2)
-            self.parent.db_write(self.VKEY3, value3)
-            self.parent.db_write(self.VKEY4, value4)
-            self.parent.db_write(self.VKEY5, value5)
-            self.parent.db_write(self.VKEY6, value6)
-            self.parent.db_write(self.VKEY7, value7)
-            self.parent.db_write(self.VKEY8, value8)
+            clkState = GPIO.input(self.clk)
+            dtState = GPIO.input(self.dt)
+            if self.btn == "True" and not (GPIO.input(self.btnpin)):
+                if clkState != self.clkLastState:
+                    if dtState != clkState:
+                        self.btnstcounter += self.btnincr
+                    else:
+                        self.btnstcounter -= self.btnincr
+                    self.parent.db_write(self.btnkey, self.btnstcounter)
+            else:
+                if clkState != self.clkLastState:
+                    if dtState != clkState:
+                        self.counter += self.incr
+                    else:
+                        self.counter -= self.incr
+                    self.parent.db_write(self.key, self.counter)
+            self.clkLastState = clkState
+    
         self.running = False
 
     def stop(self):
