+++ src/plugins/rpi_virtualwire/virtualwire/virtualwire.py
@@ -1,3 +1,4 @@
+# coding: utf8
 #!/usr/bin/env python
 
 # vw.py
